Removed ambiguity of lifecycle macro/module.
