+++ server/swim_agent/src/agent_lifecycle/on_stop.rs
