+++ server/swim_agent/src/agent_lifecycle/stateless/mod.rs
@@ -17,7 +17,7 @@ use std::marker::PhantomData;
 use static_assertions::assert_impl_all;
 use swim_api::handlers::{FnHandler, NoHandler};
 
-use crate::lifecycle::AgentLifecycle;
+use crate::agent_lifecycle::AgentLifecycle;
 
 use super::{lane_event::LaneEvent, on_start::OnStart, on_stop::OnStop};
