+++ server/swim_agent/src/agent_model/tests/fake_lifecycle.rs
@@ -16,8 +16,8 @@ use swim_model::Text;
 use tokio::sync::mpsc;
 
 use crate::{
+    agent_lifecycle::{lane_event::LaneEvent, on_start::OnStart, on_stop::OnStop},
     event_handler::{HandlerAction, StepResult},
-    lifecycle::{lane_event::LaneEvent, on_start::OnStart, on_stop::OnStop},
     meta::AgentMetadata,
 };
