+++ server/swim_agent/src/lanes/map/lifecycle/on_remove.rs
@@ -17,8 +17,8 @@ use std::collections::HashMap;
 use swim_api::handlers::{FnHandler, NoHandler};
 
 use crate::{
+    agent_lifecycle::utility::HandlerContext,
     event_handler::{EventHandler, UnitHandler},
-    lifecycle::utility::HandlerContext,
 };
 
 /// Lifecycle event for the `on_remove` event of a map lane.
