+++ server/swim_agent/src/lanes/value/lifecycle/mod.rs
@@ -16,7 +16,7 @@ use std::marker::PhantomData;
 
 use swim_api::handlers::{FnHandler, NoHandler};
 
-use crate::lifecycle::utility::HandlerContext;
+use crate::agent_lifecycle::utility::HandlerContext;
 
 use self::{
     on_event::{OnEvent, OnEventShared},
